+++ src/extensions/export/bootstrap-table-export.js
@@ -106,7 +106,7 @@ $.BootstrapTable = class extends $.BootstrapTable {
             <div class="export ${this.constants.classes.buttonsDropdown}">
             <button class="${this.constants.buttonsClass} dropdown-toggle"
             aria-label="Export"
-            data-toggle="dropdown"
+            ${this.constants.dataToggle}="dropdown"
             type="button"
             title="${o.formatExport()}">
             ${o.showButtonIcons ? Utils.sprintf(this.constants.html.icon, o.iconsPrefix, o.icons.export) : ''}